end key issue fixed, repo reorganization
